feat: infinite scroll 훅 로직 분리
